adds a base class to split functionality
